% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{glance.cens_np_fit}
\alias{glance.cens_np_fit}
\title{Glance at a non-parametric NPMLE fit}
\usage{
\method{glance}{cens_np_fit}(x, ...)
}
\arguments{
\item{x}{A \code{cens_np_fit} object.}

\item{...}{Unused.}
}
\value{
A one-row data frame.
}
\description{
One-row model-level summary.
}
\examples{
glance(desc_np(as_interval_data(c(0, 1, 5, 8), c(1, 3, 5, 8))))
}
