% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute-bounds.R
\name{is_cens_bounds}
\alias{is_cens_bounds}
\title{Test whether an object is a \code{cens_bounds}}
\usage{
is_cens_bounds(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
\code{TRUE} if \code{x} is a \code{cens_bounds} object, otherwise \code{FALSE}.
}
\description{
Test whether an object is a \code{cens_bounds}
}
