% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-cens.R
\name{is_cens_data}
\alias{is_cens_data}
\title{Test whether an object is a \code{cens_data}}
\usage{
is_cens_data(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
\code{TRUE} if \code{x} is a \code{cens_data} object, otherwise \code{FALSE}.
}
\description{
Test whether an object is a \code{cens_data}
}
