% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cens-frame.R
\name{is_cens_list}
\alias{is_cens_list}
\title{Test whether an object is a \code{cens_list}}
\usage{
is_cens_list(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
\code{TRUE} if \code{x} is a \code{cens_list}, otherwise \code{FALSE}.
}
\description{
Test whether an object is a \code{cens_list}
}
