% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc-nonparam.R
\name{plot.cens_np_fit}
\alias{plot.cens_np_fit}
\title{Plot the cumulative distribution of a non-parametric NPMLE fit}
\usage{
\method{plot}{cens_np_fit}(x, xlab = "Concentration", ylab = "Cumulative probability", ...)
}
\arguments{
\item{x}{A \code{cens_np_fit} object.}

\item{xlab, ylab}{Axis labels.}

\item{...}{Further arguments passed to \code{\link[=plot]{plot()}}.}
}
\value{
\code{x}, invisibly.
}
\description{
Draws the estimated cumulative distribution function \code{P(X <= conc)} as a step
function on the original measurement scale.
}
